+++ anymap/__init__.py
@@ -4,15 +4,13 @@ __author__ = """Qiusheng Wu"""
 __email__ = "giswqs@gmail.com"
 __version__ = "0.1.2"
 
-from .anymap import (
-    MapWidget,
-    MapLibreMap,
-    MapboxMap,
-    CesiumMap,
-    PotreeMap,
-    DeckGLMap,
-    MapCompare,
-)
+from .base import MapWidget
+from .maplibre import MapLibreMap
+from .mapbox import MapboxMap
+from .cesium import CesiumMap
+from .potree import PotreeMap
+from .deckgl import DeckGLMap
+from .compare import MapCompare
 
 __all__ = [
     "MapWidget",
