+++ anymap/base.py
@@ -1,0 +1,201 @@
+"""Base class for interactive map widgets using anywidget."""
+
+import anywidget
+import traitlets
+from typing import Dict, List, Any, Optional
+
+
+class MapWidget(anywidget.AnyWidget):
+    """Base class for interactive map widgets using anywidget."""
+
+    # Widget traits for communication with JavaScript
+    center = traitlets.List([0.0, 0.0]).tag(sync=True)
+    zoom = traitlets.Float(2.0).tag(sync=True)
+    width = traitlets.Unicode("100%").tag(sync=True)
+    height = traitlets.Unicode("600px").tag(sync=True)
+    style = traitlets.Unicode("").tag(sync=True)
+
+    # Communication traits
+    _js_calls = traitlets.List([]).tag(sync=True)
+    _js_events = traitlets.List([]).tag(sync=True)
+
+    # Internal state
+    _layers = traitlets.Dict({}).tag(sync=True)
+    _sources = traitlets.Dict({}).tag(sync=True)
+
+    def __init__(self, **kwargs):
+        """Initialize the map widget."""
+        super().__init__(**kwargs)
+        self._event_handlers = {}
+        self._js_method_counter = 0
+
+    def call_js_method(self, method_name: str, *args, **kwargs) -> None:
+        """Call a JavaScript method on the map instance."""
+        call_data = {
+            "id": self._js_method_counter,
+            "method": method_name,
+            "args": args,
+            "kwargs": kwargs,
+        }
+        self._js_method_counter += 1
+
+        # Trigger sync by creating new list
+        current_calls = list(self._js_calls)
+        current_calls.append(call_data)
+        self._js_calls = current_calls
+
+    def on_map_event(self, event_type: str, callback):
+        """Register a callback for map events."""
+        if event_type not in self._event_handlers:
+            self._event_handlers[event_type] = []
+        self._event_handlers[event_type].append(callback)
+
+    @traitlets.observe("_js_events")
+    def _handle_js_events(self, change):
+        """Handle events from JavaScript."""
+        events = change["new"]
+        for event in events:
+            event_type = event.get("type")
+            if event_type in self._event_handlers:
+                for handler in self._event_handlers[event_type]:
+                    handler(event)
+
+    def set_center(self, lat: float, lng: float) -> None:
+        """Set the map center."""
+        self.center = [lat, lng]
+
+    def set_zoom(self, zoom: float) -> None:
+        """Set the map zoom level."""
+        self.zoom = zoom
+
+    def fly_to(self, lat: float, lng: float, zoom: Optional[float] = None) -> None:
+        """Fly to a specific location."""
+        options = {"center": [lat, lng]}
+        if zoom is not None:
+            options["zoom"] = zoom
+        self.call_js_method("flyTo", options)
+
+    def add_layer(self, layer_id: str, layer_config: Dict[str, Any]) -> None:
+        """Add a layer to the map."""
+        # Store layer in local state for persistence
+        current_layers = dict(self._layers)
+        current_layers[layer_id] = layer_config
+        self._layers = current_layers
+
+        self.call_js_method("addLayer", layer_config, layer_id)
+
+    def remove_layer(self, layer_id: str) -> None:
+        """Remove a layer from the map."""
+        # Remove from local state
+        current_layers = dict(self._layers)
+        if layer_id in current_layers:
+            del current_layers[layer_id]
+            self._layers = current_layers
+
+        self.call_js_method("removeLayer", layer_id)
+
+    def add_source(self, source_id: str, source_config: Dict[str, Any]) -> None:
+        """Add a data source to the map."""
+        # Store source in local state for persistence
+        current_sources = dict(self._sources)
+        current_sources[source_id] = source_config
+        self._sources = current_sources
+
+        self.call_js_method("addSource", source_id, source_config)
+
+    def remove_source(self, source_id: str) -> None:
+        """Remove a data source from the map."""
+        # Remove from local state
+        current_sources = dict(self._sources)
+        if source_id in current_sources:
+            del current_sources[source_id]
+            self._sources = current_sources
+
+        self.call_js_method("removeSource", source_id)
+
+    def get_layers(self) -> Dict[str, Any]:
+        """Get all layers currently on the map."""
+        return dict(self._layers)
+
+    def get_sources(self) -> Dict[str, Any]:
+        """Get all sources currently on the map."""
+        return dict(self._sources)
+
+    def clear_layers(self) -> None:
+        """Clear all layers from the map."""
+        layer_ids = list(self._layers.keys())
+        for layer_id in layer_ids:
+            self.remove_layer(layer_id)
+
+    def clear_sources(self) -> None:
+        """Clear all sources from the map."""
+        source_ids = list(self._sources.keys())
+        for source_id in source_ids:
+            self.remove_source(source_id)
+
+    def clear_all(self) -> None:
+        """Clear all layers and sources from the map."""
+        self.clear_layers()
+        self.clear_sources()
+
+    def to_html(
+        self,
+        filename: Optional[str] = None,
+        title: str = "Anymap Export",
+        width: str = "100%",
+        height: str = "600px",
+        **kwargs,
+    ) -> str:
+        """Export the map to a standalone HTML file.
+
+        Args:
+            filename: Optional filename to save the HTML. If None, returns HTML string.
+            title: Title for the HTML page
+            width: Width of the map container
+            height: Height of the map container
+            **kwargs: Additional arguments passed to the HTML template
+
+        Returns:
+            HTML string content
+        """
+        # Get the current map state
+        map_state = {
+            "center": self.center,
+            "zoom": self.zoom,
+            "width": width,
+            "height": height,
+            "style": self.style,
+            "_layers": dict(self._layers),
+            "_sources": dict(self._sources),
+        }
+
+        # Add class-specific attributes
+        if hasattr(self, "map_style"):
+            map_state["map_style"] = self.map_style
+        if hasattr(self, "bearing"):
+            map_state["bearing"] = self.bearing
+        if hasattr(self, "pitch"):
+            map_state["pitch"] = self.pitch
+        if hasattr(self, "antialias"):
+            map_state["antialias"] = self.antialias
+        if hasattr(self, "access_token"):
+            map_state["access_token"] = self.access_token
+
+        # Generate HTML content
+        html_content = self._generate_html_template(map_state, title, **kwargs)
+
+        # Save to file if filename is provided
+        if filename:
+            with open(filename, "w", encoding="utf-8") as f:
+                f.write(html_content)
+
+        return html_content
+
+    def _generate_html_template(
+        self, map_state: Dict[str, Any], title: str, **kwargs
+    ) -> str:
+        """Generate the HTML template with map state.
+
+        This method should be overridden by subclasses to provide library-specific templates.
+        """
+        raise NotImplementedError("Subclasses must implement _generate_html_template")
