+++ anymap/maplibre.py
@@ -1,0 +1,285 @@
+"""MapLibre GL JS implementation of the map widget."""
+
+import pathlib
+import traitlets
+from typing import Dict, List, Any, Optional, Union
+import json
+
+from .base import MapWidget
+
+# Load MapLibre-specific js and css
+with open(pathlib.Path(__file__).parent / "static" / "maplibre_widget.js", "r") as f:
+    _esm_maplibre = f.read()
+
+with open(pathlib.Path(__file__).parent / "static" / "maplibre_widget.css", "r") as f:
+    _css_maplibre = f.read()
+
+
+class MapLibreMap(MapWidget):
+    """MapLibre GL JS implementation of the map widget."""
+
+    # MapLibre-specific traits
+    map_style = traitlets.Unicode("https://demotiles.maplibre.org/style.json").tag(
+        sync=True
+    )
+    bearing = traitlets.Float(0.0).tag(sync=True)
+    pitch = traitlets.Float(0.0).tag(sync=True)
+    antialias = traitlets.Bool(True).tag(sync=True)
+
+    # Define the JavaScript module path
+    _esm = _esm_maplibre
+    _css = _css_maplibre
+
+    def __init__(
+        self,
+        center: List[float] = [0.0, 0.0],
+        zoom: float = 2.0,
+        map_style: str = "https://demotiles.maplibre.org/style.json",
+        width: str = "100%",
+        height: str = "600px",
+        bearing: float = 0.0,
+        pitch: float = 0.0,
+        **kwargs,
+    ):
+        """Initialize MapLibre map widget.
+
+        Args:
+            center: Map center as [latitude, longitude]
+            zoom: Initial zoom level
+            map_style: MapLibre style URL or style object
+            width: Widget width
+            height: Widget height
+            bearing: Map bearing (rotation) in degrees
+            pitch: Map pitch (tilt) in degrees
+        """
+        super().__init__(
+            center=center,
+            zoom=zoom,
+            width=width,
+            height=height,
+            map_style=map_style,
+            bearing=bearing,
+            pitch=pitch,
+            **kwargs,
+        )
+
+    def set_style(self, style: Union[str, Dict[str, Any]]) -> None:
+        """Set the map style."""
+        if isinstance(style, str):
+            self.map_style = style
+        else:
+            self.call_js_method("setStyle", style)
+
+    def set_bearing(self, bearing: float) -> None:
+        """Set the map bearing (rotation)."""
+        self.bearing = bearing
+
+    def set_pitch(self, pitch: float) -> None:
+        """Set the map pitch (tilt)."""
+        self.pitch = pitch
+
+    def add_geojson_layer(
+        self,
+        layer_id: str,
+        geojson_data: Dict[str, Any],
+        layer_type: str = "fill",
+        paint: Optional[Dict[str, Any]] = None,
+    ) -> None:
+        """Add a GeoJSON layer to the map."""
+        source_id = f"{layer_id}_source"
+
+        # Add source
+        self.add_source(source_id, {"type": "geojson", "data": geojson_data})
+
+        # Add layer
+        layer_config = {"id": layer_id, "type": layer_type, "source": source_id}
+
+        if paint:
+            layer_config["paint"] = paint
+
+        self.add_layer(layer_id, layer_config)
+
+    def add_marker(self, lat: float, lng: float, popup: Optional[str] = None) -> None:
+        """Add a marker to the map."""
+        marker_data = {"coordinates": [lng, lat], "popup": popup}
+        self.call_js_method("addMarker", marker_data)
+
+    def fit_bounds(self, bounds: List[List[float]], padding: int = 50) -> None:
+        """Fit the map to given bounds."""
+        self.call_js_method("fitBounds", bounds, {"padding": padding})
+
+    def add_raster_layer(
+        self,
+        layer_id: str,
+        source_url: str,
+        paint: Optional[Dict[str, Any]] = None,
+        layout: Optional[Dict[str, Any]] = None,
+    ) -> None:
+        """Add a raster layer to the map."""
+        source_id = f"{layer_id}_source"
+
+        # Add raster source
+        self.add_source(
+            source_id, {"type": "raster", "tiles": [source_url], "tileSize": 256}
+        )
+
+        # Add raster layer
+        layer_config = {"id": layer_id, "type": "raster", "source": source_id}
+
+        if paint:
+            layer_config["paint"] = paint
+        if layout:
+            layer_config["layout"] = layout
+
+        self.add_layer(layer_id, layer_config)
+
+    def add_vector_layer(
+        self,
+        layer_id: str,
+        source_url: str,
+        source_layer: str,
+        layer_type: str = "fill",
+        paint: Optional[Dict[str, Any]] = None,
+        layout: Optional[Dict[str, Any]] = None,
+    ) -> None:
+        """Add a vector tile layer to the map."""
+        source_id = f"{layer_id}_source"
+
+        # Add vector source
+        self.add_source(source_id, {"type": "vector", "url": source_url})
+
+        # Add vector layer
+        layer_config = {
+            "id": layer_id,
+            "type": layer_type,
+            "source": source_id,
+            "source-layer": source_layer,
+        }
+
+        if paint:
+            layer_config["paint"] = paint
+        if layout:
+            layer_config["layout"] = layout
+
+        self.add_layer(layer_id, layer_config)
+
+    def add_image_layer(
+        self,
+        layer_id: str,
+        image_url: str,
+        coordinates: List[List[float]],
+        paint: Optional[Dict[str, Any]] = None,
+    ) -> None:
+        """Add an image layer to the map."""
+        source_id = f"{layer_id}_source"
+
+        # Add image source
+        self.add_source(
+            source_id, {"type": "image", "url": image_url, "coordinates": coordinates}
+        )
+
+        # Add raster layer for the image
+        layer_config = {"id": layer_id, "type": "raster", "source": source_id}
+
+        if paint:
+            layer_config["paint"] = paint
+
+        self.add_layer(layer_id, layer_config)
+
+    def _generate_html_template(
+        self, map_state: Dict[str, Any], title: str, **kwargs
+    ) -> str:
+        """Generate HTML template for MapLibre GL JS."""
+        # Serialize map state for JavaScript
+        map_state_json = json.dumps(map_state, indent=2)
+
+        html_template = f"""<!DOCTYPE html>
+<html>
+<head>
+    <title>{title}</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <script src="https://unpkg.com/maplibre-gl@5.6.1/dist/maplibre-gl.js"></script>
+    <link href="https://unpkg.com/maplibre-gl@5.6.1/dist/maplibre-gl.css" rel="stylesheet">
+    <style>
+        body {{
+            margin: 0;
+            padding: 20px;
+            font-family: Arial, sans-serif;
+        }}
+        #map {{
+            width: {map_state['width']};
+            height: {map_state['height']};
+            border: 1px solid #ccc;
+        }}
+        h1 {{
+            margin-top: 0;
+            color: #333;
+        }}
+    </style>
+</head>
+<body>
+    <h1>{title}</h1>
+    <div id="map"></div>
+
+    <script>
+        // Map state from Python
+        const mapState = {map_state_json};
+
+        // Initialize MapLibre map
+        const map = new maplibregl.Map({{
+            container: 'map',
+            style: mapState.map_style || 'https://demotiles.maplibre.org/style.json',
+            center: [mapState.center[1], mapState.center[0]], // Convert [lat, lng] to [lng, lat]
+            zoom: mapState.zoom || 2,
+            bearing: mapState.bearing || 0,
+            pitch: mapState.pitch || 0,
+            antialias: mapState.antialias !== undefined ? mapState.antialias : true
+        }});
+
+        // Restore layers and sources after map loads
+        map.on('load', function() {{
+            // Add sources first
+            const sources = mapState._sources || {{}};
+            Object.entries(sources).forEach(([sourceId, sourceConfig]) => {{
+                try {{
+                    map.addSource(sourceId, sourceConfig);
+                }} catch (error) {{
+                    console.warn(`Failed to add source ${{sourceId}}:`, error);
+                }}
+            }});
+
+            // Then add layers
+            const layers = mapState._layers || {{}};
+            Object.entries(layers).forEach(([layerId, layerConfig]) => {{
+                try {{
+                    map.addLayer(layerConfig);
+                }} catch (error) {{
+                    console.warn(`Failed to add layer ${{layerId}}:`, error);
+                }}
+            }});
+        }});
+
+        // Add navigation controls
+        map.addControl(new maplibregl.NavigationControl());
+
+        // Add scale control
+        map.addControl(new maplibregl.ScaleControl());
+
+        // Log map events for debugging
+        map.on('click', function(e) {{
+            console.log('Map clicked at:', e.lngLat);
+        }});
+
+        map.on('load', function() {{
+            console.log('Map loaded successfully');
+        }});
+
+        map.on('error', function(e) {{
+            console.error('Map error:', e);
+        }});
+    </script>
+</body>
+</html>"""
+
+        return html_template
